requirements.txt
fastapi
uvicorn
transformers
diffusers
safetensors
sentence-transformers
faiss-cpu
pandas
numpy
accelerate
